Add subnet connection and change subscription to VM.
We are no more able to ssh on a VM public IP because port 22 is blocked.
The idea is to create the VM on a subnet which is accessed through VPN.
Then, you are able to ssh on the VM private IP.

Note that, the VM must be in the same region than the subnet.

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>